refactor: adding ci testcase for unused counters
