+++ integration/test/dashboard_json_test.go
@@ -23,6 +23,7 @@ const (
 )
 
 var restDataCollectors = []string{"Rest"}
+var zapiDataCollectors = []string{"Zapi"}
 
 var fileSet []string
 
@@ -249,7 +250,7 @@ func counterIsMissing(flavor string, counter string, waitFor time.Duration) bool
 	}
 	query := counter + `{datacenter=~"` + strings.Join(restDataCollectors, "|") + `"}`
 	if flavor == zapi {
-		query = counter + `{datacenter!~"` + strings.Join(restDataCollectors, "|") + `"}`
+		query = counter + `{datacenter!~"` + strings.Join(zapiDataCollectors, "|") + `"}`
 	}
 	return !hasDataInDB(query, waitFor)
 }
